<ver2 1.01 업데이트 1. 특정 값에서의 계산오류 수정 > Integer.parseInt를 Double.parseDouble로 수정함.
2. 숫자 단위별로 , 삽입함.

+++ app/src/main/java/com/hyunseo/donghaechoi/maplestoryexpmanager/Calculater_EXP.java
@@ -136,7 +136,7 @@ public class Calculater_EXP extends AppCompatActivity implements View.OnClickLis
 
                     // 현재 Level까지의 모든 경험치를 더한 값에
                     // 현재 경험치 값을 더함
-                    sumNowExp += Integer.parseInt(mNowExp.getText().toString());
+                    sumNowExp += Double.parseDouble(mNowExp.getText().toString());
 
                     // 백분율을 구하는 식
                     double calculation1 = (sumNowExp / sumAimExp) * 100;
